Have a separate module for real time cloud, implements #110
