+++ src/main/java/com/digitalpebble/spruce/modules/realtimecloud/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package provides access to data from the GSF Realtime cloud project
+ *
+ * @see <a href="Source: https://github.com/Green-Software-Foundation/real-time-cloud>Real Time Cloud</a>
+ */
+package com.digitalpebble.spruce.modules.realtimecloud;
+
