+++ src/main/resources/gsf_realtime_cloud/region-mappings.csv
